udp reader implemented
