+++ hspytools/__init__.py
@@ -1,3 +1,1 @@
-from hspytools.hdf5_mgr import hdf5_mgr
-from hspytools.tparray import TPArray
-from hspytools.helpers import HTPAdGUI_FileReader,HTPA_ByteStream_Reader,QuadriPolygon
+
