+++ hspytools/helpers.py
@@ -23,445 +23,7 @@ import re
 
 from .tparray import TPArray
 
-class HTPAdGUI_FileReader():
-    
-    def __init__(self,width,height):
-        
-        self.width = width
-        self.height = height
-        
-        
-        # Depending on the sensor size, the content of the bds-file is
-        # organized as follows
-        self.tparray = TPArray(width,height)
-        # data_order = ArrayType.get_serial_data_order()
-        
-        
-    def read_htpa_video(self,path):
-        
-        # Check if path is provided as pathlib
-        if not isinstance(path,Path):
-            path = Path(path)
-        
-        # get suffix
-        suffix = path.suffix
 
-        if suffix.casefold() == ('.txt').casefold():
-            df_video, header = self._import_txt(path) 
-        elif suffix.casefold() == ('.bds').casefold():
-            df_video, header = self._import_bds(path)
-        else:
-            print('File extension not recognized.')
-            return None
-        
-        return df_video, header
-    
-    def _import_txt(self,path,**kwargs):
-        
-        sep = kwargs.pop('sep',' ')
-        skiprows = kwargs.pop('skiprows',1)
-        
-        # Get columns names
-        columns = self.tparray.get_serial_data_order()
-        
-        # Read txt with pandas to DataFrame
-        txt_content = pd.read_csv(path,
-                                  sep=sep,
-                                  skiprows=skiprows,
-                                  header=None)
-        
-        # txt file contains a time stamp at the end, that's not in the bds-file
-        # use only the first columns that contain actual data
-        txt_content = txt_content[np.arange(0,len(columns),1)]
-        
-        # rename columns appropriately
-        txt_content.columns = columns
-
-        # rename index appropriately
-        txt_content.index = range(0,len(txt_content))
-        txt_content.index.name = 'image_id'
-        
-        header = None
-        
-        return txt_content, header
-    
-    def _import_bds(self,bds_path,**kwargs):
-        
-        # open file and save content byte by byte in list
-        bds_content = []
-        
-        with open(bds_path, "rb") as f:
-            
-            # Read the header byte by byte until '\n'
-            header = []
-            header_end = False
-            
-            while header_end == False:
-                header.append(f.read(1))
-
-                if header[-1].decode()=='\n':
-                    header_end = True
-            
-            # Join bytes of header together
-            header = bytes().join(header)
-
-                
-            # Read two bytes at a time 
-            while (LSb := f.read(2)):
-                
-                # and combine in LSb fashion
-                bds_content.append(int.from_bytes(LSb, 
-                                                  byteorder='little'))
-               
-        # Cast the data to a DataFrame of appropriate size
-        columns = self.tparray.get_serial_data_order()
-        
-        # If the last frame has not been fully transmitted, reshaping fails
-        # Therefore throw out the last incomplete frame
-        num_full_frames = int(len(bds_content)/len(columns))
-        bds_content = bds_content[0:num_full_frames*len(columns)]
-        bds_content = (np.array(bds_content)).reshape(-1,len(columns))
-        bds_content = pd.DataFrame(data=bds_content,
-                                   columns = columns)
-        
-        bds_content.index = range(len(bds_content))
-        bds_content.index.name = 'image_id'
-        
-        return bds_content, header
-        
-    def reverse(self,df_video):
-        """
-        Function for rotating a video by 180°. Intended for postprocessing 
-        a video sequence from a sensor that was mounted upside-down
-
-        Parameters
-        ----------
-        df_video : pd.DataFrame
-            DataFrame containing the content of a text or bds file exported 
-            from the HTPAdGUI
-
-        Returns
-        -------
-        df_video: pd.DataFrame
-            Upside-down version of the original video.
-
-        """
-        
-        # Rotation must be applied to the pixel values and the electrical offsets
-        pix_cols = self.tparray._pix
-        e_off_cols = self.tparray._e_off
-        
-        NROFBLOCKS = self.tparray._DevConst['NROFBLOCKS']
-        width = self.tparray._width
-        height = self.tparray._height
-        
-        # go through the DataFrame image by image
-        for i in df_video.index:
-            
-            # get pixel and electrical offsets values
-            pix_val = df_video.loc[i,pix_cols]
-            e_off_val = df_video.loc[i,e_off_cols]
-            
-            # reshape values according to sensor
-            pix_val = pix_val.values.reshape(self.tparray._npsize)
-            e_off_val = e_off_val.values.reshape((int(height/NROFBLOCKS),
-                                                  width))
-            
-            # rotate both arrays by 180°
-            pix_val = np.rot90(pix_val,k=2)
-            e_off_val = np.rot90(e_off_val,k=2)
-            
-            # reshape to a row vector and write to dataframe
-            df_video.loc[i,pix_cols] = pix_val.flatten()
-            df_video.loc[i,e_off_cols] = e_off_val.flatten()
-            
-        # Return the rotated dataframe
-        return df_video
-    
-    def export_bds(self,df_video,header,bds_path,**kwargs):
-        """
-        Export a video sequence in dataframe format to a .bds file that is
-        compatible with HTPAdGUI
-        
-
-        Parameters
-        ----------
-        df_video : pd.DataFrame
-            DESCRIPTION.
-        header : byte
-            DESCRIPTION.
-        path : pathlib.Path
-            DESCRIPTION.
-
-        Returns
-        -------
-        None.
-
-        """
-        
-        mode = kwargs.pop('mode','x')
-        
-        
-        # Check if file already exists
-        if bds_path.exists():
-            if mode=='x':
-                print('File exists. No data will be written. Pass mode="x" to overwrite.')
-            elif mode=='w':
-                print('File exists and will be overwritten.')
-                os.remove(bds_path)
-        
-        # first write the header to the list byte by byte
-        bds_content = []
-        bds_content.append(header)
-        # Go over the video sequence image by image, convert all integers to 
-        # bytes and append to the list
-        
-        for i in df_video.index:
-            
-            # get the whole row as a row vector
-            row = df_video.loc[i].values
-            
-            # cast every integer to a byte in little endian byteorder
-            for val in row:
-                bds_content.append(int(val).to_bytes(length=2,byteorder='little'))
-            
-        # Write bytes to file
-        with open(bds_path, "wb") as bds_file:
-            [bds_file.write(b) for b in bds_content]
-            
-        return None
-    
-    def export_png(self,df_video,path):
-        """
-        A function for writing a video sequence given as a DataFrame to .png
-        frame by frame in a sepcified folder (path)
-        
-
-        Parameters
-        ----------
-        df_video : pd.DataFrame
-            DESCRIPTION.
-
-        Returns
-        -------
-        None.
-
-        """
-        
-        # Get shape of sensor array
-        npsize = self.tparray._npsize
-        
-        # Get columns with pixel values
-        pix_cols = self.tparray._pix
-        
-        # Get rid of everything else
-        df_video = df_video[pix_cols]
-        
-       
-       
-       
-        if not path.exists():
-            path.mkdir(parents=True,exist_ok=False)
-        
-        for i in df_video.index:
-            img = df_video.loc[[i]].values.reshape(npsize)
-
-            
-            file_name = str(i) + '.png'
-            
-            img = self._scale_img(img)
-
-            matplotlib.image.imsave(path / file_name, img)
-        
-        return None
-
-    # def export_avi(self,df_video,video_name,path,**kwargs):
-    #     """
-    #     A function for writing a video sequence given as a DataFrame to .avi
-    #     in a sepcified folder (path)
-        
-
-    #     Parameters
-    #     ----------
-    #     df_video : pd.DataFrame
-    #         DESCRIPTION.
-
-    #     Returns
-    #     -------
-    #     None.
-
-    #     """        
-
-    #     # Framerate
-    #     fps = kwargs.pop('fps',8)
-        
-    #     # Get shape of sensor array
-    #     size = self.tparray._size
-    #     npsize = self.tparray._npsize
-        
-    #     # Get columns with pixel values
-    #     pix_cols = self.tparray._pix
-        
-    #     # Get rid of everything else
-    #     df_video = df_video[pix_cols]
-        
-    #     if not path.exists():
-    #         path.mkdir(parents=True,exist_ok=False)
-        
-    #     # Initialize video writer
-    #     writer = imageio.get_writer((path / (video_name + '.avi')).as_posix(),
-    #                                 fps=fps,
-    #                                 macro_block_size=1)
-        
-    #     # codec = cv2.VideoWriter_fourcc(*'MJPG')
-    #     # codec = cv2.VideoWriter_fourcc(*'H264')
-        
-    #     # video = cv2.VideoWriter((path / (video_name + '.avi')).as_posix(),
-    #     #                         codec,
-    #     #                         fs,
-    #     #                         size,
-    #     #                         isColor = True)  
-                
-    #     # Get colormap
-    #     cmap = matplotlib.cm.get_cmap('plasma')      
-        
-    #     # Loop over all frames and write to mp4
-    #     for i in df_video.index:
-            
-    #         img = df_video.loc[i].values.reshape(npsize)
-            
-    #         # Normalization 
-    #         img = ( img - img.min() ) / (img.max() - img.min())
-            
-    #         # Apply colormap
-    #         RGBA = (cmap(img)*255).astype('uint8')
-            
-    #         # opencv
-    #         # BGR = cv2.cvtColor(RGBA, cv2.COLOR_RGB2BGR)
-    #         # video.write(BGR)
-            
-    #         # imageio_ffmpeg
-    #         writer.append_data(RGBA)
-
-    #     # opencv
-    #     # cv2.destroyAllWindows()
-    #     # video.release()
-        
-    #     # imageio_ffmpeg
-    #     writer.close()
-       
-    #     return None   
-    
-    def _scale_img(self,img):
-        
-        # Get width an height of image
-        w = self.tparray._width
-        h = self.tparray._height
-        
-        # Crop image by 10 % circumferential
-        crop_w = int(np.ceil(0.1*w))
-        crop_h = int(np.ceil(0.1*h))
-        
-        # Crop the image by 3 pixels to get rid of corners.
-        img_crop = img[crop_h:h-crop_h,crop_w:w-crop_w]
-        
-        # Use the pixel values in the cropped frame to scale the image
-        dK_max  = img_crop.max()
-        dK_min  = img_crop.min()
-        
-        img = ( img - dK_min ) / (dK_max - dK_min)
-        
-        img[img<=0] = 0
-
-
-        return img
-    
-    def _flip(self,df_video):
-
-        w = self.width
-        h = self.height
-        
-        pix_cols = self.tparray._pix
-        
-        for i in df_video.index:
-            img = df_video.loc[i,pix_cols].values.reshape((h,w))
-            img = np.flip(img,axis=1).flatten()
-            df_video.loc[i,pix_cols] = img
-
-            
-        return df_video
-               
- 
-class HTPA_ByteStream_Reader():
-    def __init__(self,width,height):
-        
-        self.width = width
-        self.height = height
-        
-        self.tparray = TPArray(width,height)
-        
-        # Initialize an array to which to write data
-        self.data_cols = self.tparray.get_serial_data_order()
-        self.data = np.zeros((len(self.data_cols)))
-    
-    def bytes_to_np(self,byte_stream:list):
-        
-        if not isinstance(byte_stream, list):
-            TypeError('byte_stream needs to be a list of bytes')
-        
-        # Get the column headers for this array type
-        data_cols = self.tparray.get_serial_data_order()
-        
-        # Zero data array
-        self.data.fill(0)
-        
-        j=0
-        
-        # Loop over all elements / packages in list
-        for package in byte_stream:
-            
-            # Read the first byte, it's the package index
-            package_index = package[0]
-            
-            # Loop over all bytes and combine MSB and LSB
-            idx = np.arange(1,len(package),2)
-            
-            for i in idx:    
-                self.data[j] = int.from_bytes(package[i:i+2], byteorder='little')
-                j = j+1
-       
-        Warning("Check if pixels are in appropriate order (i.e. if Bodo's)"+\
-                "and pyplots/opencvs coordinate system are the same")
-        
-        return self.data
-    
-    def bytes_to_df(self,byte_stream:list):
-        
-        self.bytes_to_np(byte_stream)
-        
-        df_data = pd.DataFrame(data=[self.data],
-                               columns=self.data_cols)
-        
-        return df_data
-        
-        
-    # def bytes_to_img(self,byte_stream):
-        
-             
-    #     # Loop over all bytes and combine MSB and LSB
-    #     idx = np.arange(0,len(byte_stream),2)
-        
-    #     img = np.zeros((1,self.width*self.height))
-    #     j=0
-        
-    #     for i in idx:    
-    #         img[0,j] = int.from_bytes(byte_stream[i:i+2], byteorder='little')
-    #         j = j+1
-            
-    #     img = img.reshape((self.height,self.width))
-    #     img = np.flip(img,axis=1)
-        
-    #     return img
         
 
 class QuadriPolygon():
@@ -1091,629 +653,9 @@ def pd_interpolate_MI (df_input, df_toInterpolate,col):
     return pd.concat([df_input, df_toInterpolate]).sort_index()
 
 
-class LuT:
-    
-    def __init__(self,**kwargs):
-        
-        pass
-    
-    def LuT_from_df(self,df):
-        
-        
-        
-        
-        self.LuT = df
-        
-    def LuT_from_csv(self,csv_path,offset):
-        
-        self.offset = offset
-        
-        # Import data
-        LuT = pd.read_csv(csv_path, sep=',',header=0,index_col = 0)
-        
-        # Convert column header to int
-        LuT.columns = np.array([int(c) for c in LuT.columns])
-        
-        # Subtract offset
-        LuT.index = LuT.index - offset
-        
-        self.LuT = LuT
-        
-    def LuT_from_xlsx(self,xlsx_path,sheet_name,**kwargs):
-        
-        index_col = kwargs.pop('index_col',0)
-        usecols = kwargs.pop('usecols','A,D:O')
-        skiprows = kwargs.pop('skiprows',12)
-        header = kwargs.pop('header',0)
-        # dtype = kwargs.pop('dtype','object')
-        
-        df = pd.read_excel(xlsx_path,
-                           sheet_name = sheet_name,
-                           skiprows = skiprows,
-                           index_col = index_col,
-                           usecols = usecols,
-                           header = header)
-        
-        # Delete all commata
-        df = df.replace(',','',regex=True)
-        
-        # Cast all columns to int
-        df = df.astype(np.int32)
-        
-        # Rename index
-        df.index.name = 'Ud'
-        
-        # That's it
-        self.LuT = df
-        
-        return None
-    
-    def LuT_to_xlsx(self,xlsx_path):
-        
-        # Load LuT
-        LuT = self.LuT.copy()
-        
-        # Initialize writer object
-        writer = self._init_xlswriter(xlsx_path)
-        
-        # Start a list where columns are in the right order
-        columns_ordered = list(LuT.columns)
-        
-        # Reindex so Ud becomes column
-        LuT = LuT.reset_index(drop=False)        
-        
-        # Add a column vector for the offset-free voltage signal
-        LuT['Ud_norm'] = LuT['Ud'] - LuT['Ud'].min() 
-        
-        # Opening and closing brackets
-        LuT['br_o'] = ['{']*len(LuT)
-        LuT['br_c'] = ['},']*len(LuT)
-        
-        columns_ordered = ['Ud','Ud_norm','br_o'] + columns_ordered + ['br_c']
-        
-        LuT = LuT[columns_ordered]
-        
-        LuT.to_excel(writer,
-                     sheet_name = xlsx_path.stem,
-                     startrow=12,
-                     index=False)
-        
-        # Then write normalized Ud to row 9
-        df_V = pd.DataFrame(data=[],
-                            columns=np.arange(0,len(LuT),1,dtype=int),
-                            index=['V'])
-        
-        columns_ordered = list(df_V.columns)
-                
-        df_V.loc['V'] = LuT['Ud_norm'].values.astype(int)
-        
-        df_V = df_V.astype('str')
-        df_V[df_V.columns[:-1]] = df_V[df_V.columns[:-1]]+','
-        
-        df_V['br_o'] = '{'
-        df_V['br_c'] = '};'
-
-        columns_ordered =  ['br_o'] + columns_ordered + ['br_c']
-        
-        df_V = df_V[columns_ordered]
-        
-        df_V.to_excel(writer,
-                      sheet_name = xlsx_path.stem,
-                      startrow=9,
-                      index=True,
-                      header=False)
-        
-        # and Ta to row 10
-        df_Ta = pd.DataFrame(data=[],
-                            columns=list(self.LuT.columns),
-                            index=['Ta'])
-        
-        columns_ordered = list(df_Ta.columns)
-                
-        df_Ta.loc['Ta'] = list(self.LuT.columns)
-        
-
-        
-        df_Ta = df_Ta.astype('str')
-        df_Ta[df_Ta.columns[:-1]] = df_Ta[df_Ta.columns[:-1]]+','
-
-        
-        df_Ta['br_o'] = '{'
-        df_Ta['br_c'] = '};'
-
-        columns_ordered =  ['br_o'] + columns_ordered + ['br_c']
-        
-        df_Ta = df_Ta[columns_ordered]
-        
-        df_Ta.to_excel(writer,
-                       sheet_name = xlsx_path.stem,
-                       startrow=10,
-                       index=True,
-                       header=False)
-        
-        writer.save()
-            
-    def LuT_from_HTPAxls(self,sheet_name,**kwargs):
-              
-        xlsx_path = Path('T:/Projekte/HTPA8x8_16x16_32x31/Datasheet/LookUpTablesHTPA.xlsm')
-        
-        self.LuT_from_xlsx(xlsx_path,sheet_name,**kwargs)
-        
-        return None
-        
-    def inverse_eval_LuT(self,data,Ta_col='Tamb0',To_col='To_meas'):
-        
-        # Check if index is unique, otherwise loop will exract more than one
-        # measurement per loop and algorithm brakes down
-        if not data.index.is_unique:
-            print('Data index is not unique. reset_index() is applied!')
-            data = data.reset_index()
-        
-        
-        for meas in data.index:
-            
-            LuT_copy = self.LuT.copy()
-        
-            Ta_meas = data.loc[meas,Ta_col]
-            To_meas = data.loc[meas,To_col]
-            
-            # find the "last" column in old LuT that is smaller than the 
-            # measured Ta 
-            col_idx = LuT_copy.columns < Ta_meas
-            LuT_col = LuT_copy.columns[col_idx][-1]
-            
-            # get neighbouring column
-            Ta_col_n = LuT_copy.columns[LuT_copy.columns.get_loc(LuT_col)+1]
-            
-            # create a new column by interpolation
-            new_col = int(np.round(Ta_meas))
-            f = (new_col-LuT_col) / (Ta_col_n-LuT_col)
-            LuT_copy[new_col] = LuT_copy[LuT_col] + \
-                f*(LuT_copy[Ta_col_n]-LuT_copy[LuT_col])
-            
-            # Find index of las To in that column that is smaller than the measured To
-            Ud_row = LuT_copy.loc[LuT_copy[new_col]<To_meas].index[-1]
-            
-            # get neighbouring row
-            Ud_row_n = LuT_copy.index[LuT_copy.index.get_loc(Ud_row)+1]
-            
-            # Calculate Ud_est
-            dT = LuT_copy.loc[Ud_row_n,new_col] - LuT_copy.loc[Ud_row,new_col]
-            dU = Ud_row_n-Ud_row
-            
-            f =  dU/dT
-            
-            data.loc[meas,'Ud_LuT'] = Ud_row + \
-                (To_meas - LuT_copy.loc[Ud_row,new_col]) * f
-                
-        return data
-        
-    def eval_LuT(self,data,Ta_col='Tamb0',Ud_col='Ud'):
-        
-        LuT = self.LuT
-        
-        for meas in data.index:
-    
-            Ta_meas = data.loc[meas,Ta_col]
-            Ud = data.loc[meas,Ud_col]
-        
-            # Find columns and indeces for bilinear interpolation
-            col_idx = LuT.columns < Ta_meas
-            LuT_col = LuT.columns[col_idx][-1]
-            Ta_col_n = LuT.columns[LuT.columns.get_loc(LuT_col)+1]
-            
-            row_idx = LuT.index < Ud
-            Ud_row = LuT.index[row_idx][-1]
-            Ud_row_n = LuT.index[LuT.index.get_loc(Ud_row)+1]
-            
-            
-            rect_pnts = LuT.loc[Ud_row:Ud_row_n,[LuT_col,Ta_col_n]]
-            
-            data.loc[meas,'To_LuT'] = \
-                self._get_To(rect_pnts,Ta_meas,Ud)
-                
-        return data
-        
-        
-    def _get_To(self,points,Ta_meas,Ud_meas):
-        
-        x0 = points.columns[0]
-        x1 = points.columns[1]
-        
-        y0 = points.index[0]
-        y1 = points.index[1]
-    
-        
-        pt1 = (x0,y0,points.loc[y0,x0])
-        pt2 = (x0,y1,points.loc[y1,x0])
-        pt3 = (x1,y0,points.loc[y0,x1])
-        pt4 = (x1,y1,points.loc[y1,x1])
-            
-        rect_pnts = np.array([pt1,pt2,pt3,pt4])
-        
-        return self._bilinear_interpolation(Ta_meas,Ud_meas,rect_pnts)
-        
-    def _bilinear_interpolation(self,x, y, points):
-        '''Interpolate (x,y) from values associated with four points.
-    
-        The four points are a list of four triplets:  (x, y, value).
-        The four points can be in any order.  They should form a rectangle.
-    
-            >>> bilinear_interpolation(12, 5.5,
-            ...                        [(10, 4, 100),
-            ...                         (20, 4, 200),
-            ...                         (10, 6, 150),
-            ...                         (20, 6, 300)])
-            165.0
-    
-        '''
-        # See formula at:  http://en.wikipedia.org/wiki/Bilinear_interpolation
-    
-        # points = sorted(points)               # order points by x, then by y
-        (x1, y1, q11), (_x1, y2, q12), (x2, _y1, q21), (_x2, _y2, q22) = points
-    
-        if x1 != _x1 or x2 != _x2 or y1 != _y1 or y2 != _y2:
-            raise ValueError('points do not form a rectangle')
-        if not x1 <= x <= x2 or not y1 <= y <= y2:
-            raise ValueError('(x, y) not within the rectangle')
-    
-        return (q11 * (x2 - x) * (y2 - y) +
-                q21 * (x - x1) * (y2 - y) +
-                q12 * (x2 - x) * (y - y1) +
-                q22 * (x - x1) * (y - y1)
-               ) / ((x2 - x1) * (y2 - y1) + 0.0)
-        
-    def plot_LuT(self):
-        
-        LuT = self.LuT
-        
-        fig = plt.figure()
-        ax = fig.add_subplot(111, projection='3d')
-        
-        X = np.array(LuT.columns).reshape((1,-1))
-        X = np.repeat(X,len(LuT),axis=0)
-        
-        Y = np.array(LuT.index).reshape((-1,1))
-        Y = np.repeat(Y,len(LuT.columns),axis=1)
-        
-        Z = LuT.values
-        
-                
-        ax.plot_surface(X,Y,Z,cmap=matplotlib.cm.coolwarm,antialiased=False)
-        ax.set_xlabel('Tamb0')
-        ax.set_ylabel('Ud')
-        ax.set_zlabel('To_pred')
-        
-        pass
-            
-    def _init_xlswriter(self,path):
-        
-        # Load or create workbook
-        writer = pd.ExcelWriter(path, engine='openpyxl')
-        writer.book.create_sheet(title=path.stem)
-        
-        return writer
-    
-class HTPA_UDP_Reader():
-    
-    def __init__(self,width,height):
-        
-        self.width = width
-        self.height = height
-
-        # Initialize a TPArray object, which contains all information regarding
-        # how data is stored and organized for this array type
-        self._tparray = TPArray(width,height)
-    
-        self._port = 30444
-        
-        # Create a DataFrame to store all bound devices in
-        self.col_dict = {'IP':str,'MAC-ID':str,'Arraytype':int,
-                         'status':str}
-        self.index = pd.Index(data=[],dtype=int,name='DevID')
-        
-        self._devices = pd.DataFrame(data=[],
-                                    columns = self.col_dict.keys(),
-                                    index = self.index)
-        
-        self._sockets = {}
-    
-        # Initialize ByteStream Reader for convertings bytes to pandas
-        # dataframes
-        self.bytestream_reader = HTPA_ByteStream_Reader(width,height)
-        
-    
-    @property
-    def port(self):
-        return self._port
-    @property
-    def tparray(self):
-        return self._tparray
-    
-    @property
-    def devices(self):
-        return self._devices
-    @devices.setter
-    def devices(self,df):
-        devices_old = self._devices
-        self._devices = pd.concat(devices_old,df)
-
-    @property
-    def sockets(self):
-        return self._sockets
-    @sockets.setter
-    def sockets(self,socket:dict):
-        self._sockets.update(socket)
 
     
-    def broadcast(self):
-        Warning('Not implemented yet.')
-        return None
-    
-    def bind_tparray(self,ip:str):
-        """
-        Creates a socket for the device with the given ip
 
-        Parameters
-        ----------
-        ip : int
-            DESCRIPTION.
-
-        Returns
-        -------
-        None.
-
-        """
-        # Create the udp socket
-        udp_socket = socket.socket(socket.AF_INET,          # Internet
-                                   socket.SOCK_DGRAM)       # UDP
-        
-        # Create server address
-        server_address = (ip,self._port)
-        
-        # Set timeout to 1 second
-        socket.setdefaulttimeout(1)
-        
-        # Try calling the device and check if it's a HTPA device
-        try:
-            _ = udp_socket.sendto(bytes('Calling HTPA series devices',
-                                        'utf-8'),
-                                       server_address)
-            call = udp_socket.recv(self.tparray._package_size)
-        except:
-            Exception('Calling HTPA series device failed')
-            return None
-        
-        # If calling was successfull, extract basic information from the 
-        # answer string
-        dev_info = self._callstring_to_information(call)
-        
-        # Next try to bind the device
-        try:
-            _ = udp_socket.sendto(bytes('Bind HTPA series device',
-                                        'utf-8'),
-                                       server_address)
-            bind = udp_socket.recv(self.tparray._package_size)
-        except:
-            Exception('Calling HTPA series device failed')
-            return None
-        
-        dev_info['status'] = 'bound'
-        
-        # Add socket to dictionary
-        dev_id = dev_info.index.item()
-        self.sockets = {dev_id:udp_socket}
-        
-        # Append new device to device list and store
-        self._devices = dev_info
-        
-        return self.devices.copy()
-
-    def release_tparray(self,dev_id):
-        
-        # Get device information from the device list
-        dev_info = self.devices.loc[[dev_id]]
-        
-        # If more than one devices have the same device id, return error
-        if len(dev_info) != 1:
-            Exception('Multiple devices have the same device id.')
-            
-            # Get udp socket
-            udp_socket = self.sockets[dev_id]
-            
-            # Create the udp socket
-            # udp_socket = socket.socket(socket.AF_INET,          # Internet
-            #                            socket.SOCK_DGRAM)       # UDP
-            
-            # Set timeout to 1 second
-            # socket.setdefaulttimeout(1)
-            
-            # Create server address
-            server_address = (dev_info['IP'].item(), self.port)
-            
-            # Send message to device to stop streaming bytes
-            _ = udp_socket.sendto(bytes('x','utf-8'),server_address)
-    
-    def start_continuous_bytestream(self,dev_id):
-        
-        # Get device information from the device list
-        dev_info = self.devices.loc[[dev_id]]
-        
-        # If more than one devices have the same device id, return error
-        if len(dev_info) != 1:
-            Exception('Multiple devices have the same device id.')
-        
-        # Get udp socket
-        udp_socket = self.sockets[dev_id]
-        
-        # Create the udp socket
-        # udp_socket = socket.socket(socket.AF_INET,          # Internet
-        #                            socket.SOCK_DGRAM)       # UDP
-        
-        # Set timeout to 1 second
-        # socket.setdefaulttimeout(1)
-        
-        # Create server address
-        server_address = (dev_info['IP'].item(), self.port)
-        
-        # Connect socket to address
-        # connect = udp_socket.connect(server_address)
-        
-        # Send message to device to start streaming bytes
-        _ = udp_socket.sendto(bytes('K','utf-8'),server_address)
-
-    def stop_continuous_bytestream(self,dev_id):
-        
-        # Get device information from the device list
-        dev_info = self.devices.loc[[dev_id]]
-        
-        # If more than one devices have the same device id, return error
-        if len(dev_info) != 1:
-            Exception('Multiple devices have the same device id.')
-        
-        # Get udp socket
-        udp_socket = self.sockets[dev_id]
-        
-        # Create the udp socket
-        # udp_socket = socket.socket(socket.AF_INET,          # Internet
-        #                            socket.SOCK_DGRAM)       # UDP
-        
-        # Set timeout to 1 second
-        # socket.setdefaulttimeout(1)
-        
-        # Create server address
-        server_address = (dev_info['IP'].item(), self.port)
-        
-        # Connect socket to address
-        # udp_socket.connect(server_address)
-        
-        # Send message to device to start streaming bytes
-        _ = udp_socket.sendto(bytes('X','utf-8'),server_address)        
-        
-        return None
-    
-    def read_continuous_bytestream(self,dev_id):
-        
-        # Get device information from the device list
-        dev_info = self.devices.loc[[dev_id]]
-        
-        # If more than one devices have the same device id, return error
-        if len(dev_info) != 1:
-            Exception('Multiple devices have the same device id.')
-        
-        # Get udp socket
-        udp_socket = self.sockets[dev_id]
-        
-        # Create the udp socket
-        # udp_socket = socket.socket(socket.AF_INET,          # Internet
-        #                            socket.SOCK_DGRAM)       # UDP
-        
-        # Set timeout to 1 second
-        # socket.setdefaulttimeout(1)
-        
-        # Create server address
-        server_address = (dev_info['IP'].item(), self.port)
-        
-        # _ = udp_socket.sendto(bytes('K','utf-8'),server_address)
-        
-        # Connect socket to address
-        # connect = udp_socket.connect(('192.168.240.144', 58495))
-        
-
-        
-        # Create variable that indicates if DataFrame was constructed success-
-        # fully
-        success = False
-        while success == False:
-            
-            # Initialize list for received packages            
-            packages = []
-            
-            # Read incoming packages until one with the package index 1 is received
-            sync = False
-            
-            while sync == False:
-                
-                # Receive package
-                package = udp_socket.recv(self.tparray._package_size)
-                
-                # Get index of package
-                package_index = int(package[0])
-                
-                # Check if it is equal to 1
-                if package_index == 1:
-                    packages.append(package)
-                    sync = True
-            
-            for p in range(2,self.tparray._package_num+1):
-                
-                # Receive package
-                package = udp_socket.recv(self.tparray._package_size)
-                
-                # Get index of package
-                package_index = int(package[0])
-                
-                # Check if package index has the expected value
-                if package_index == p:
-                    packages.append(package)
-                
-            # In the end check if as many packages were received as expected
-            if len(packages) == self.tparray._package_num:
-                
-                # If yes, pass the packages to the class that parses the
-                # bytes into a pandas dataframe
-                # df_frame = self.bytestream_reader.bytes_to_df(packages)
-                
-                df_frame = self.bytestream_reader.bytes_to_np(packages)
-                
-                
-                success = True
-                
-            else:
-                print('Frame lost.')
-                sync = False
-
-                
-  
-
-        return df_frame
-        
-    def _callstring_to_information(self,call:bytes):
-        """
-        Extracts information from the ridiculously long and inefficient answer
-        to the device call.
-
-        Parameters
-        ----------
-        call : bytes
-            DESCRIPTION.
-
-        Returns
-        -------
-        dev_info : TYPE
-            DESCRIPTION.
-
-        """
-
-        call = call.decode('utf-8')
-        
-        # Extract information from callstring
-        arraytype = int(call.split('Arraytype')[1].split('MODTYPE')[0])
-        mac_id = re.findall(r'\d{2}\.\d{2}\.\w{2}\.\d{2}.\d{2}.\d{2}',call)[0]
-        ip = re.findall(r'\d{3}\.\d{3}\.\d{3}.\d{3}',call)[0]
-        dev_id = int(call.split('DevID:')[1].split('Emission')[0])
-        
-                
-        dev_info = pd.DataFrame(data = [],
-                                columns = self.col_dict.keys(),
-                                index = self.index)
-        
-        dev_info.loc[dev_id,['IP','MAC-ID','Arraytype']] = \
-            [ip,mac_id,arraytype]
-        
-        dev_info = dev_info.astype(self.col_dict)    
-        
-        return dev_info
